十二桁まで対応

+++ src/main/kotlin/number_names/NumberToKanji.kt
@@ -8,29 +8,29 @@ class NumberToKanji {
 
     companion object {
         val SINGLE_NUMBER_KANJI_MAP = mapOf(
-                0 to "",
-                1 to "一",
-                2 to "二",
-                3 to "三",
-                4 to "四",
-                5 to "五",
-                6 to "六",
-                7 to "七",
-                8 to "八",
-                9 to "九"
+                0L to "",
+                1L to "一",
+                2L to "二",
+                3L to "三",
+                4L to "四",
+                5L to "五",
+                6L to "六",
+                7L to "七",
+                8L to "八",
+                9L to "九"
         )
 
         val HEADER_NUMBER_KANJI_MAP = mapOf(
-                0 to "",
-                1 to "",
-                2 to "二",
-                3 to "三",
-                4 to "四",
-                5 to "五",
-                6 to "六",
-                7 to "七",
-                8 to "八",
-                9 to "九"
+                0L to "",
+                1L to "",
+                2L to "二",
+                3L to "三",
+                4L to "四",
+                5L to "五",
+                6L to "六",
+                7L to "七",
+                8L to "八",
+                9L to "九"
         )
 
         val CYCLE_DIGITS_KANJI_MAP = mapOf(
@@ -45,7 +45,7 @@ class NumberToKanji {
         )
     }
 
-    fun convert(number: Int) :String? {
+    fun convert(number: Long) :String? {
         for (largeDigitsKanjiEntry in LARGE_DEGITS_KANJI_MAP) {
             if (number >= largeDigitsKanjiEntry.key) {
                 return convert(number / largeDigitsKanjiEntry.key) + largeDigitsKanjiEntry.value + convert(number % largeDigitsKanjiEntry.key)
